Rename agent-input-system-prompt.ts to agent-prompt.ts
Shorten the file name.
