+++ apps/dashboard/src/server/usecases/agent.test.ts
@@ -5,7 +5,7 @@ vi.mock("../infras/local-hermeum-config", () => ({ LocalConfig: vi.fn() }));
 vi.mock("../infras/ai-sdk", () => ({ AiSdkGenerator: vi.fn() }));
 
 import { AgentUseCase } from "./agent";
-import { AGENT_INPUT_SYSTEM_PROMPT } from "./agent-input-system-prompt";
+import { AGENT_INPUT_SYSTEM_PROMPT } from "./agent-prompt";
 import type { AiGenerator } from "./adaptors/generator";
 import type { ConfigAdaptor } from "./adaptors/config";
 import type { Runtime } from "./adaptors/runtime";
