+++ apps/dashboard/src/server/usecases/agent.ts
@@ -20,7 +20,7 @@ import {
 import { AiSdkGenerator } from "../infras/ai-sdk";
 import { KubernetesClient } from "../infras/kubernetes/client";
 import { LocalConfig } from "../infras/local-hermeum-config";
-import { AGENT_INPUT_SYSTEM_PROMPT } from "./agent-input-system-prompt";
+import { AGENT_INPUT_SYSTEM_PROMPT } from "./agent-prompt";
 import { AiGenerator } from "./adaptors/generator";
 import { Runtime } from "./adaptors/runtime";
 import { ConfigAdaptor } from "./adaptors/config";
